Moved from diagnostic/status to node/health
