+++ rest/src/routes/diagnosticRoutes.js
@@ -22,15 +22,9 @@ const routeResultTypes = require('./routeResultTypes');
 const routeUtils = require('./routeUtils');
 const { version: sdkVersion } = require('../../../catapult-sdk/package.json');
 const { version: restVersion } = require('../../package.json');
-const nodeInfoCodec = require('../sockets/nodeInfoCodec');
-const catapult = require('catapult-sdk');
-
-const packetHeader = catapult.packet.header;
-const { PacketType } = catapult.packet;
-const { BinaryParser } = catapult.parser;
 
 module.exports = {
-	register: (server, db, services) => {
+	register: (server, db) => {
 		server.get('/diagnostic/blocks/:height/limit/:limit', (req, res, next) => {
 			const parseUint = paramName => routeUtils.parseArgument(req.params, paramName, 'uint');
 			const height = parseUint('height');
@@ -60,44 +54,5 @@ module.exports = {
 				res.send({ payload: storageInfo, type: routeResultTypes.storageInfo });
 				next();
 			}));
-
-		server.get('/diagnostic/status', (req, res, next) => {
-			const tryParseNodeInfoPacket = packet => {
-				try {
-					const binaryParser = new BinaryParser();
-					binaryParser.push(packet.payload);
-					return nodeInfoCodec.deserialize(binaryParser);
-				} catch (error) {
-					return undefined;
-				}
-			};
-
-			const okMessage = 'OK';
-			const downMessage = 'down';
-
-			// Check database status
-			const dbStatus = true === db.database.serverConfig.isConnected() ? okMessage : downMessage;
-
-			// Check apiNode status
-			const packetBuffer = packetHeader.createBuffer(PacketType.nodeDiscoveryPullPing, packetHeader.size);
-			const apiNodeStatus = services.connections.singleUse()
-				.then(connection => connection.pushPull(packetBuffer, services.config.apiNode.timeout))
-				.then(packet => (undefined !== tryParseNodeInfoPacket(packet) ? okMessage : downMessage))
-				.catch(e => e.message);
-
-			return apiNodeStatus.then(status => {
-				res.send({
-					payload: {
-						statusInfo: {
-							apiNode: status,
-							db: dbStatus
-						}
-					},
-					type: routeResultTypes.statusInfo
-				});
-
-				return next();
-			});
-		});
 	}
 };
