+++ rest/test/routes/allRoutes_spec.js
@@ -60,14 +60,14 @@ describe('all routes', () => {
 			'/network',
 			'/node/info',
 			'/node/time',
+			'/node/health',
 
 			'/transaction/:transactionId',
 			'/transaction/:hash/status',
 
 			'/diagnostic/blocks/:height/limit/:limit',
 			'/diagnostic/server',
-			'/diagnostic/storage',
-			'/diagnostic/status'
+			'/diagnostic/storage'
 		]);
 	});
